Update internal types and add 'safe' conversion macros
The internal type casting can get a little messy, as mentioned in #2.
It is difficult to not use `usize` directly for some things, due to the
nature they are used (accessing arrays a lot).

For now, we only make minor modifications and replace all casts with
macros which can be adjusted in the future to provide safer casting
(i.e. no negative i32 casts to usize, assert!(x > 0)).

+++ src/controller.rs
@@ -37,7 +37,7 @@ impl CLike for Action {
 #[derive(Default)]
 pub struct Controller {
     /// The length each action has occured for in ticks.
-    pub time: [usize; 8],
+    pub time: [u64; 8],
 
     /// Which actions are currently active.
     pub active: [bool; 8]
@@ -65,7 +65,7 @@ impl Controller {
     }
 
     /// Query how long an action has been active for.
-    pub fn time(&self, action: Action) -> usize {
+    pub fn time(&self, action: Action) -> u64 {
         self.time[action.to_usize()]
     }
